--- src/app/api/products/route.ts
@@ -1,19 +1,0 @@
-import { NextRequest, NextResponse } from "next/server";
-
-import { prisma } from "@/services/prisma";
-// import data from '../../../services/mock.json';
-
-export async function GET(request: NextRequest) {
-  const product = await prisma.product.findMany({
-    include: {
-      files: true,
-      categories: true,
-    }
-  });
-
-  if (!product) {
-    throw new Error('Product Server Error');
-  };
-
-  return NextResponse.json(product);
-};
