+++ src/app/panel/panel/index.tsx
@@ -1,8 +1,6 @@
 "use client"
 
 import React, { useState, Fragment, useContext } from 'react';
-import { revalidateTag } from 'next/cache';
-import Image from 'next/image';
 
 import { DotLottiePlayer } from '@dotlottie/react-player';
 import { useTheme } from 'styled-components';
@@ -13,7 +11,8 @@ import { Product, Button } from '@/components';
 import { Box, Block, Success, Tag, TextAlignLeft } from '@/assets/svg/icons';
 import { NotificationContext } from '@/hooks/notification';
 
-import { Container, Modal, UploadMore, Navigation, ProductEmpty } from './styles';
+import { revalidatePanelProduct } from '../actions';
+import { Container, UploadMore, Navigation, ProductEmpty } from './styles';
 
 type Props = {
   products: ProductsProps[]
@@ -23,15 +22,9 @@ const Panel: React.FC<Props> = ({ products }) => {
   const theme = useTheme();
 
   const { setNotification } = useContext(NotificationContext);
-  const [product, setProduct] = useState<ProductsProps | null>(null);
   
-  const handleDelete = (prod: ProductsProps) => {
-    setProduct(prod);
-    document.body.style.overflow = "hidden";
-  };
-
-  const handleClose = async () => {
-    const res = await fetch(`/api/product?id=${product?.id}`, {
+  const handleDelete = async (prod: ProductsProps) => {
+    const res = await fetch(`/api/product?id=${prod?.id}`, {
       method: 'DELETE'
     });
 
@@ -39,16 +32,8 @@ const Panel: React.FC<Props> = ({ products }) => {
       return setNotification({ icon: Block, type: 'failed', message: 'Failed to delete', active: `${Math.random()}_show` });
     };
 
-    setProduct(null);
-    revalidateTag('products');
-    document.body.style.overflowY = "scroll";
-
+    revalidatePanelProduct();
     return setNotification({ icon: Success, type: 'success', message: 'Deleted successful', active: `${Math.random()}_show` });
-  };
-
-  const styles = {
-    opacity: product? 1 : 0,
-    zIndex: product? 6 : -1,
   };
 
   const lottie_styles  = {
@@ -81,24 +66,6 @@ const Panel: React.FC<Props> = ({ products }) => {
           <TextAlignLeft width={20} height={20} stroke={theme.colors.primary} strokeWidth={1.8} />
         </UploadMore>
       </Navigation>
-
-      <Modal style={styles}>
-        {product?.id && (
-          <Image
-            src={product?.files[0]?.url}
-            width={300}
-            height={300}
-            style={{
-              borderRadius: 3
-            }}
-            alt='delete'
-          />
-        )}
-
-        <div>
-          <Button onClick={handleClose}>Delete</Button>
-        </div>
-      </Modal>
     </Fragment>
   )
 };
