+++ src/app/panel/product/page.tsx
@@ -1,45 +1,6 @@
-import { ProductsProps, CategoryProps } from '@/global/interfaces';
-
 import Redit from './redit';
 
-async function getCategories(): Promise<CategoryProps[]> {
-  const res = await fetch(`${process.env.NEXT_URL}/api/category`, {
-    method: 'GET',
-    cache: 'no-store',
-    headers: {
-      'Content-Type': 'application/json'
-    }
-  });
-
-  if (!res.ok) {
-    throw new Error('Product by Id')
-  };
-
-  const data = await res.json();
-
-  return data;
-};
-
-async function getProductById(id: string): Promise<ProductsProps | null> {
-  if (!id) return null;
-
-  const res = await fetch(`${process.env.NEXT_URL}/api/product?id=${id}`, {
-    method: 'GET',
-    // cache: 'no-cache',
-    cache: 'no-store',
-    headers: {
-      'Content-Type': 'application/json'
-    }
-  });
-
-  if (!res.ok) {
-    throw new Error('Product by Id');
-  };
-
-  const data = await res.json();
-
-  return data;
-};
+import { getCategories, getProductById } from '../actions';
 
 export default async function Page(params: any) {
   const categories = await getCategories();
